server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.1
python-multipart>=0.0.6
pydantic>=2.0.0
email-validator>=2.0.0
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlmodel>=0.0.24
pyjwt>=2.0.0
